devop: rename use wagmi to handle connecting to wallets instead
